# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentation of UPay API.",
        "title": "UPay API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/admin/login": {
            "post": {
                "description": "login admin to get jwt token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Login Admin.",
                "parameters": [
                    {
                        "description": "JSON username and user_pass",
                        "name": "admin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.NewAdminLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.LoginAdminSuccess"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product_types": {
            "post": {
                "description": "create new product type with name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Type"
                ],
                "summary": "Create New Product Type.",
                "parameters": [
                    {
                        "description": "JSON name",
                        "name": "product_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docs.CreateProductType"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/docs.CreateProductTypeSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/docs.CreateProductTypeFail"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "docs.CreateProductType": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Pulsa"
                }
            }
        },
        "docs.CreateProductTypeFail": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "string",
                    "example": "error insert product type"
                }
            }
        },
        "docs.CreateProductTypeSuccess": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "string",
                    "example": "success"
                },
                "product_type_name": {
                    "type": "string",
                    "example": "Pulsa"
                }
            }
        },
        "docs.LoginAdminSuccess": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "string",
                    "example": "success"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2NTQyNzA2OTAsImlkIjoxLCJ1c2VybmFtZSI6ImFkbWluIn0.9Qif8_Ug6Uy_oxDXIuIui3nepCPFc1jJ6mO6wEhiuHE"
                }
            }
        },
        "docs.NewAdminLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "admin"
                },
                "username": {
                    "type": "string",
                    "example": "admin123"
                }
            }
        }
    }
}